% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{ActiveIdent}
\alias{ActiveIdent}
\alias{ActiveIdent<-}
\title{ActiveIdent}
\usage{
ActiveIdent(object)

ActiveIdent(object) <- value
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{value}{identity source, e.g. \code{colLabels} or a \code{colData} column name}
}
\value{
current active identity source or updated object
}
\description{
get or set the active identity field
}
