% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{AggregateExpression}
\alias{AggregateExpression}
\title{AggregateExpression}
\usage{
AggregateExpression(object, group_by, assay.type = "counts", fun = "sum")
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{group_by}{A character vector of column names in colData(object) to group by.
Typically c("cluster", "sample").}

\item{assay.type}{Assay to aggregate. Default is "counts".}

\item{fun}{Aggregation function. Default is "sum".}
}
\value{
A SummarizedExperiment object with aggregated counts.
The colData contains the unique combinations of the grouping variables.
}
\description{
Aggregate expression for pseudobulk analysis
}
\details{
Aggregates expression values by groups (e.g. cluster and sample) to create a pseudobulk object.
}
