% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{as.SCE}
\alias{as.SCE}
\title{as.SCE}
\usage{
as.SCE(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Additional arguments passed to Seurat::as.SingleCellExperiment}
}
\value{
A SingleCellExperiment object
}
\description{
Convert to SingleCellExperiment object
}
\details{
Convert a Seurat object to a SingleCellExperiment object.
}
