% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{as.Seurat}
\alias{as.Seurat}
\title{as.Seurat}
\usage{
as.Seurat(object, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{...}{Additional arguments passed to Seurat::as.Seurat}
}
\value{
A Seurat object
}
\description{
Convert to Seurat object
}
\details{
Convert a SingleCellExperiment object to a Seurat object.
}
