% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{CommandLog}
\alias{CommandLog}
\title{CommandLog}
\usage{
CommandLog(object, details = FALSE)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{details}{logical, whether to include raw \code{params} and \code{outputs} list columns}
}
\value{
a data.frame summarizing recorded commands
}
\description{
inspect sclet command log
}
