% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\title{DefaultAssay}
\usage{
DefaultAssay(object)

DefaultAssay(object) <- value
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{value}{assay name}
}
\value{
current default assay or updated object
}
\description{
get or set the default assay
}
