% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{defaultDbNames}
\alias{defaultDbNames}
\title{Documentation for the data
Default databases by organism}
\format{
Named character vector
}
\source{
GEO number
}
\usage{
defaultDbNames
}
\description{
Documentation for the data
Default databases by organism
}
\keyword{datasets}
