% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{DefaultGraph}
\alias{DefaultGraph}
\alias{DefaultGraph<-}
\title{DefaultGraph}
\usage{
DefaultGraph(object)

DefaultGraph(object) <- value
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{value}{graph name}
}
\value{
current default graph or updated object
}
\description{
get or set the default graph
}
