% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{DefaultLayer}
\alias{DefaultLayer}
\alias{DefaultLayer<-}
\title{DefaultLayer}
\usage{
DefaultLayer(object)

DefaultLayer(object) <- value
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{value}{layer name}
}
\value{
current default layer or updated object
}
\description{
get or set the default layer
}
