% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{DefaultReduction}
\alias{DefaultReduction}
\alias{DefaultReduction<-}
\title{DefaultReduction}
\usage{
DefaultReduction(object)

DefaultReduction(object) <- value
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{value}{reduction name}
}
\value{
current default reduction or updated object
}
\description{
get or set the default dimensional reduction
}
