% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{DimPlot}
\alias{DimPlot}
\title{DimPlot}
\usage{
DimPlot(object, reduction = "UMAP", ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{reduction}{reduction method, default is "UMAP"}

\item{...}{additional parameters passed to ggsc::sc_dim}
}
\value{
ggplot object
}
\description{
Dimensional reduction plot
}
