% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EnrichmentPlot}
\alias{EnrichmentPlot}
\title{EnrichmentPlot}
\usage{
EnrichmentPlot(object, id = NULL, showCategory = 10, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{id}{Analysis record id. If NULL, uses the active enrichment state.}

\item{showCategory}{Number of categories to show}

\item{...}{additional parameters passed to enrichplot::dotplot}
}
\value{
ggplot object
}
\description{
Enrichment Plot
}
\details{
Plot enrichment results using dotplot from enrichplot.
}
