% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{FeatureScatter}
\alias{FeatureScatter}
\title{FeatureScatter}
\usage{
FeatureScatter(object, feature1, feature2)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{feature1}{selected feature}

\item{feature2}{selected feature}
}
\value{
a scatter plot
}
\description{
scatter plot to compare 2 features
}
