% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{FindMarkers_pseudobulk}
\alias{FindMarkers_pseudobulk}
\title{FindMarkers_pseudobulk}
\usage{
FindMarkers_pseudobulk(object, design, contrast, ...)
}
\arguments{
\item{object}{A SingleCellExperiment or SummarizedExperiment object (output of AggregateExpression)}

\item{design}{A formula for the design matrix (e.g. ~ condition).}

\item{contrast}{A character vector specifying the contrast (e.g. c("condition", "treated", "control")).}

\item{...}{Additional arguments passed to DESeq2::DESeq}
}
\value{
A data.frame with DE results
}
\description{
Find markers using pseudobulk DE (DESeq2)
}
\details{
Performs differential expression analysis on pseudobulk data using DESeq2.
}
