% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindNeighbors}
\alias{FindNeighbors}
\title{FindNeighbors}
\usage{
FindNeighbors(object, dims, k = 10)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{dims}{number of dimensions to be used to build the KNN graph}

\item{k}{number of neighbors}
}
\value{
updated SingleCellExperiment object with KNN graph
}
\description{
build KNN graph
}
