% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-summary-table.R
\name{gene_summary_table}
\alias{gene_summary_table}
\title{gene_summary_table
This function append gene summary information to the input gene data frame and
open the result in various formats.}
\usage{
gene_summary_table(
  gene,
  gene_col = "gene",
  orgDb = "org.Hs.eg.db",
  keyType = "SYMBOL",
  cluster_col,
  output = "html",
  browse = TRUE
)
}
\arguments{
\item{gene}{a data frame of genes (e.g., marker genes, DE genes)}

\item{gene_col}{column of the 'gene' that contains gene IDs}

\item{orgDb}{OrgDb package name used for ID conversion when \code{keyType} is not \code{ENTREZID}}

\item{keyType}{ID type of the gene IDs}

\item{cluster_col}{column to cluster genes (optional)}

\item{output}{one of 'html', 'pdf', 'png', 'md', 'csv' or 'data.frame'}

\item{browse}{whether open the output file}
}
\value{
output file path
}
\description{
gene_summary_table
This function append gene summary information to the input gene data frame and
open the result in various formats.
}
\examples{
\dontrun{
n <- 3
topn <- pbmc.markers |>
    group_by(cluster) |>
    dplyr::filter(avg_log2FC > 1) |>
    slice_head(n = n) |>
    ungroup()

gene_summary_table(topn, gene_col = 'gene', keyType='SYMBOL')

gene_summary_table(topn, gene_col = 'gene', cluster_col = 'cluster', keyType='SYMBOL')
}
}
