% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{get_cci}
\alias{get_cci}
\title{get_cci}
\usage{
get_cci(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional communication record id. If NULL, use the active communication record.}
}
\value{
a data.frame containing standardized CCI interactions, or \code{NULL}
}
\description{
Access standard Cell-Cell Communication (CCI) interactions
}
