% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_detest}
\alias{get_detest}
\title{get_detest}
\usage{
get_detest(object, id = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{id}{Analysis record id. If NULL, uses the active detest state.}
}
\value{
The detest state record
}
\description{
Get Differential Expression Test Results
}
