% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{get_enrichment}
\alias{get_enrichment}
\title{get_enrichment}
\usage{
get_enrichment(object, id = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{id}{Analysis record id. If NULL, uses the active enrichment state.}
}
\value{
The enrichment state record
}
\description{
Get Enrichment Analysis Results
}
