% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_annotation}
\alias{has_annotation}
\title{has_annotation}
\usage{
has_annotation(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional annotation record id}
}
\value{
logical
}
\description{
Check whether annotation results are available
}
