% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_batch}
\alias{has_batch}
\title{has_batch}
\usage{
has_batch(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether batch correction results are available
}
