% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_cci}
\alias{has_cci}
\title{has_cci}
\usage{
has_cci(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional communication record id}
}
\value{
logical
}
\description{
Check whether standardized CCI results are available
}
