% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_cellchat}
\alias{has_cellchat}
\title{has_cellchat}
\usage{
has_cellchat(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether CellChat results are available
}
