% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_cellphonedb}
\alias{has_cellphonedb}
\title{has_cellphonedb}
\usage{
has_cellphonedb(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional communication record id}
}
\value{
logical
}
\description{
Check whether CellPhoneDB results are available
}
