% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_cellrank}
\alias{has_cellrank}
\title{has_cellrank}
\usage{
has_cellrank(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether CellRank results are available
}
