% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_colocalization}
\alias{has_colocalization}
\title{has_colocalization}
\usage{
has_colocalization(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether colocalization results are available
}
