% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{has_detest}
\alias{has_detest}
\title{has_detest}
\usage{
has_detest(object, id = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{id}{Analysis record id. If NULL, uses the active detest state.}
}
\value{
Logical indicating if the detest state exists
}
\description{
Check if Differential Expression Test Results exist
}
