% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{has_enrichment}
\alias{has_enrichment}
\title{has_enrichment}
\usage{
has_enrichment(object, id = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{id}{Analysis record id. If NULL, uses the active enrichment state.}
}
\value{
Logical indicating if the enrichment state exists
}
\description{
Check if Enrichment Analysis Results exist
}
