% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_geneset_scoring}
\alias{has_geneset_scoring}
\title{has_geneset_scoring}
\usage{
has_geneset_scoring(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional geneset_scoring record id}
}
\value{
logical
}
\description{
Check whether gene set scoring results are available
}
