% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_graph}
\alias{has_graph}
\title{has_graph}
\usage{
has_graph(object, name = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{name}{graph name; if NULL, use the active graph}
}
\value{
logical
}
\description{
Check whether graph results are available
}
