% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_hvg}
\alias{has_hvg}
\title{has_hvg}
\usage{
has_hvg(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether HVG metadata are available
}
