% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_integration}
\alias{has_integration}
\title{has_integration}
\usage{
has_integration(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional integration record id}
}
\value{
logical
}
\description{
Check whether integration results are available
}
