% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_mapping}
\alias{has_mapping}
\title{has_mapping}
\usage{
has_mapping(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional mapping record id}
}
\value{
logical
}
\description{
Check whether mapping results are available
}
