% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_milo}
\alias{has_milo}
\title{has_milo}
\usage{
has_milo(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional milo record id}
}
\value{
logical
}
\description{
Check whether milo results are available
}
