% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal.R
\name{has_multimodal_adt}
\alias{has_multimodal_adt}
\title{Check whether an ADT modality is present}
\usage{
has_multimodal_adt(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}
}
\value{
logical. TRUE if \code{altExp(sce, "ADT")} exists.
}
\description{
Check whether an ADT modality is present
}
