% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal.R
\name{has_multimodal_atac}
\alias{has_multimodal_atac}
\title{Check whether an ATAC modality is present}
\usage{
has_multimodal_atac(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}
}
\value{
logical. TRUE if \code{altExp(sce, "ATAC")} exists.
}
\description{
Check whether an ATAC modality is present
}
