% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_nichenet}
\alias{has_nichenet}
\title{has_nichenet}
\usage{
has_nichenet(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional communication record id}
}
\value{
logical
}
\description{
Check whether NicheNet results are available
}
