% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_perturbation}
\alias{has_perturbation}
\title{has_perturbation}
\usage{
has_perturbation(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether CellOracle Perturbation results are available
}
