% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_perturbation_priority}
\alias{has_perturbation_priority}
\title{has_perturbation_priority}
\usage{
has_perturbation_priority(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether perturbation priority results are available
}
