% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_rare_cells}
\alias{has_rare_cells}
\title{has_rare_cells}
\usage{
has_rare_cells(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether rare cell detection results are available
}
