% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_scenic}
\alias{has_scenic}
\title{has_scenic}
\usage{
has_scenic(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional scenic record id}
}
\value{
logical
}
\description{
Check whether SCENIC results are available
}
