% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_spatial}
\alias{has_spatial}
\title{has_spatial}
\usage{
has_spatial(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether Spatial Deconvolution results are available
}
