% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_spatial_niche}
\alias{has_spatial_niche}
\title{has_spatial_niche}
\usage{
has_spatial_niche(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
logical
}
\description{
Check whether spatial niche results are available
}
