% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_trajectory}
\alias{has_trajectory}
\title{has_trajectory}
\usage{
has_trajectory(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional trajectory record id}
}
\value{
logical
}
\description{
Check whether trajectory results are available
}
