% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_velocity}
\alias{has_velocity}
\title{has_velocity}
\usage{
has_velocity(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional velocity record id}
}
\value{
logical
}
\description{
Check whether velocity results are available
}
