% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{Idents}
\alias{Idents}
\title{Idents}
\usage{
Idents(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
clustering memberships
}
\description{
access the clustering memberships
}
