% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{LayerData}
\alias{LayerData}
\title{LayerData}
\usage{
LayerData(object, layer = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{layer}{layer name. If NULL, use \code{DefaultLayer(object)}.}
}
\value{
assay matrix associated with the selected layer
}
\description{
access layer data
}
