% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{Layers}
\alias{Layers}
\title{Layers}
\usage{
Layers(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
character vector of available layer names
}
\description{
list available layers
}
