% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_summary.R
\name{PipelineSummary}
\alias{PipelineSummary}
\title{PipelineSummary}
\usage{
PipelineSummary(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
a textual summary of the pipeline
}
\description{
Print a textual summary of the analysis pipeline
}
