% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cci.R
\name{plot_cci_sigmoid}
\alias{plot_cci_sigmoid}
\title{plot_cci_sigmoid}
\usage{
plot_cci_sigmoid(object, id = NULL, pval_thresh = 0.05, top_n = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{id}{The communication record id to use.}

\item{pval_thresh}{The p-value threshold for filtering interactions.}

\item{top_n}{Integer. If provided, limits the plot to the top N interactions (by probability).}
}
\value{
A ggplot object
}
\description{
Plot Cell-Cell Communication Sigmoid Diagram
}
