% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ProjectionPlot}
\alias{ProjectionPlot}
\alias{plot_projection}
\title{ProjectionPlot}
\usage{
ProjectionPlot(query, ref, reduction = "UMAP", group.by = NULL, ...)

plot_projection(query, ref, reduction = "UMAP", group.by = NULL, ...)
}
\arguments{
\item{query}{A SingleCellExperiment object representing the query dataset}

\item{ref}{A SingleCellExperiment object representing the reference dataset}

\item{reduction}{The name of the dimensional reduction to use (must be present in both objects)}

\item{group.by}{The metadata column name containing the labels. If NULL, tries to use the active identity or mapping state from query.}

\item{...}{additional parameters passed to ggplot2}
}
\value{
ggplot object
}
\description{
Projection Plot for Reference Mapping
}
\details{
Plot reference and query cells in the same dimensional reduction space.
}
