% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{QCMetrics}
\alias{QCMetrics}
\title{QCMetrics}
\usage{
QCMetrics(object)
}
\arguments{
\item{object}{a 'SingleCellExperiment' object}
}
\value{
update object with two QC metrics ('nFeature_RNA' and 'nCount_RNA')
}
\description{
add QC metrics
}
