% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{ReadH5AD}
\alias{ReadH5AD}
\title{ReadH5AD}
\usage{
ReadH5AD(file, ...)
}
\arguments{
\item{file}{Path to the .h5ad file}

\item{...}{Additional arguments passed to zellkonverter::readH5AD}
}
\value{
A SingleCellExperiment object
}
\description{
Read H5AD file
}
\details{
Wrapper for zellkonverter::readH5AD to import AnnData objects as SingleCellExperiment.
}
