% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot_list}
\alias{colData}
\alias{rowData}
\alias{colData<-}
\alias{rowData<-}
\alias{assay}
\alias{plotColData}
\alias{logNormCounts}
\alias{perCellQCMetrics}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aplot}{\code{\link[aplot]{plot_list}}}

  \item{scater}{\code{\link[scater]{plotColData}}}

  \item{scuttle}{\code{\link[scuttle]{logNormCounts}}, \code{\link[scuttle]{perCellQCMetrics}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData<-}}}
}}

