% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{RenameIdents}
\alias{RenameIdents}
\title{RenameIdents}
\usage{
RenameIdents(object, new_ids)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{new_ids}{new ID labels}
}
\value{
updated SingleCellExperiment object
}
\description{
rename cluster ids
}
