% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{RunCellPhoneDB}
\alias{RunCellPhoneDB}
\title{RunCellPhoneDB}
\usage{
RunCellPhoneDB(
  sce,
  group = NULL,
  assay_name = NULL,
  layer = NULL,
  species = "human",
  name = "cellphonedb",
  threads = 4,
  iterations = 1000,
  pvalue = 0.05,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{group}{Grouping variable used for communication analysis. If \code{NULL},
sclet resolves it from \code{ActiveIdent(sce)}.}

\item{assay_name}{the name of assay. If NULL, sclet resolves it from the selected \code{layer}.}

\item{layer}{layer used for communication analysis. If NULL, use \code{DefaultLayer(sce)}.}

\item{species}{one of human, mouse. CellPhoneDB is primarily designed for human, but sclet can map mouse genes.}

\item{name}{communication record id. Defaults to \code{"cellphonedb"}.}

\item{threads}{Number of threads to use. Defaults to 4.}

\item{iterations}{Number of iterations for statistical analysis. Defaults to 1000.}

\item{pvalue}{P-value threshold. Defaults to 0.05.}

\item{...}{Additional parameters.}
}
\value{
A SingleCellExperiment object with updated CCI state.
}
\description{
Run CellPhoneDB communication analysis
}
