% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{RunDecontX}
\alias{RunDecontX}
\title{RunDecontX}
\usage{
RunDecontX(sce, assay = "counts", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{assay}{the name of the assay to use as input. Defaults to "counts".}

\item{...}{additional parameters passed to \code{celda::decontX}}
}
\value{
an updated SingleCellExperiment object with 'decontXcounts' assay
}
\description{
Run decontX to remove ambient RNA contamination
}
