% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{RunDEtest}
\alias{RunDEtest}
\title{RunDEtest}
\usage{
RunDEtest(object, ident.1 = NULL, ident.2 = NULL, name = "detest", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{ident.1}{Identity class to define markers for. If NULL, runs FindAllMarkers.}

\item{ident.2}{A second identity class for comparison}

\item{name}{Analysis record id. Defaults to "detest".}

\item{...}{Passed to FindMarkers or FindAllMarkers}
}
\value{
A SingleCellExperiment object with detest state added
}
\description{
Run Differential Expression Test
}
