% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{RunDiffusionMap}
\alias{RunDiffusionMap}
\title{RunDiffusionMap}
\usage{
RunDiffusionMap(sce, n_components = 50, reduction = NULL, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{n_components}{number of diffusion components to compute. Defaults to 50.}

\item{reduction}{the name of the dimensional reduction to use as input (e.g. 'PCA'). Defaults to NULL (uses the active reduction).}

\item{...}{additional arguments passed to \code{destiny::DiffusionMap}}
}
\value{
an updated SingleCellExperiment object with a 'DM' reduction
}
\description{
Run Diffusion Map for complex trajectory inference
}
