% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{RunDoubletFinder}
\alias{RunDoubletFinder}
\title{RunDoubletFinder}
\usage{
RunDoubletFinder(sce, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{...}{additional parameters passed to \code{scDblFinder::scDblFinder}}
}
\value{
an updated SingleCellExperiment object with doublet scores and classes
}
\description{
Run DoubletFinder to identify and filter doublets
}
