% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{RunExplorer}
\alias{RunExplorer}
\title{RunExplorer}
\usage{
RunExplorer(object, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{...}{Additional arguments passed to iSEE::iSEE}
}
\value{
A Shiny app object
}
\description{
Run Interactive Explorer
}
\details{
Launch an iSEE interactive application for exploring the SingleCellExperiment object.
}
