% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{RunImputation}
\alias{RunImputation}
\title{RunImputation}
\usage{
RunImputation(sce, method = "ALRA", assay = "logcounts", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{method}{the imputation method to use. Currently only supports "ALRA" (zero-preserving, SVD-based imputation).}

\item{assay}{the assay to use as input. Defaults to "logcounts".}

\item{...}{additional parameters passed to the underlying imputation method (e.g., \code{ALRA::alra}).}
}
\value{
an updated SingleCellExperiment object with imputed data in a new assay (e.g., 'alra_imputed')
}
\description{
Run imputation to denoise single-cell data
}
