% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{RunNicheNet}
\alias{RunNicheNet}
\title{RunNicheNet}
\usage{
RunNicheNet(
  sce,
  group = NULL,
  receiver,
  sender = "all",
  genes_oi = NULL,
  ligand_target_matrix,
  lr_network,
  weighted_networks,
  assay_name = NULL,
  layer = NULL,
  name = "nichenet",
  top_n_ligands = 20,
  exprs_pct = 0.1
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{group}{Grouping variable used for communication analysis. If \code{NULL}, sclet resolves it from \code{ActiveIdent(sce)}.}

\item{receiver}{Character, name of the receiver cell group.}

\item{sender}{Character or character vector, name(s) of sender cell groups. Defaults to \code{"all"}.}

\item{genes_oi}{Character vector of genes of interest in the receiver. If \code{NULL}, the function will use \code{scran::findMarkers} to find DEGs for the receiver cluster vs others.}

\item{ligand_target_matrix}{NicheNet ligand-target matrix. Required.}

\item{lr_network}{NicheNet ligand-receptor network. Required.}

\item{weighted_networks}{NicheNet weighted networks. Required.}

\item{assay_name}{the name of assay. If NULL, sclet resolves it from the selected \code{layer}.}

\item{layer}{layer used for communication analysis. If NULL, use \code{DefaultLayer(sce)}.}

\item{name}{communication record id. Defaults to \code{"nichenet"}.}

\item{top_n_ligands}{Number of top ligands to extract. Defaults to 20.}

\item{exprs_pct}{Numeric, threshold for percentage of expressing cells to consider a gene expressed. Defaults to 0.10.}
}
\value{
A SingleCellExperiment object with updated CCI state.
}
\description{
Run NicheNet communication analysis
}
