% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{RunPhenotypeAssociation}
\alias{RunPhenotypeAssociation}
\title{RunPhenotypeAssociation}
\usage{
RunPhenotypeAssociation(sce, bulk_matrix, phenotype_data, family = "cox", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{bulk_matrix}{Bulk expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{phenotype_data}{Phenotype annotation of each bulk sample. It can be a continuous variable, a binary group vector, or clinical survival data (two-column matrix with 'time' and 'status').}

\item{family}{Response type for the regression model. Can be 'gaussian', 'binomial', or 'cox'.}

\item{...}{Additional arguments passed to \code{scPAS::scPAS}}
}
\value{
an updated SingleCellExperiment object with phenotype association scores added to colData
}
\description{
Run phenotype association analysis using scPAS
}
