% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell.r
\name{RunSuperCell}
\alias{RunSuperCell}
\title{RunSuperCell}
\usage{
RunSuperCell(
  object,
  assay = "logcounts",
  nHVG = 2000,
  hvg_method = "scran",
  cellname = "Barcode",
  gamma = 20,
  k.knn = 5
)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{assay}{selected assay used to build KNN graph}

\item{nHVG}{number of HVGs to use}

\item{hvg_method}{one of 'seurat' or 'scran', see also \code{FindVariableFeatures()}}

\item{cellname}{IDs to label the cells, use 'Barcode' by default}

\item{gamma}{graining level, that is number_of_cells / number_of_metacells}

\item{k.knn}{number of nearest neighbors to build KNN graph}
}
\value{
SingleCellExperiment object
}
\description{
wraper function to run SuperCell
}
