% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cci.R
\name{sc_cci_chord}
\alias{sc_cci_chord}
\title{sc_cci_chord}
\usage{
sc_cci_chord(object, signaling = NULL, id = NULL, pval_thresh = 0.05, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{signaling}{A character vector of pathways or interactions to include.}

\item{id}{The communication record id to use.}

\item{pval_thresh}{The p-value threshold for filtering interactions.}

\item{...}{Additional arguments passed to \code{circlize::chordDiagram}.}
}
\value{
NULL (Plots the diagram)
}
\description{
Plot Cell-Cell Communication Chord Diagram
}
