% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cci.R
\name{sc_cci_network}
\alias{sc_cci_network}
\title{sc_cci_network}
\usage{
sc_cci_network(
  object,
  signaling = NULL,
  id = NULL,
  pval_thresh = 0.05,
  layout = "circle",
  edge_width_scale = 1
)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{signaling}{A character vector of pathways or interactions to include.}

\item{id}{The communication record id to use.}

\item{pval_thresh}{The p-value threshold for filtering interactions.}

\item{layout}{The layout algorithm to use for the network (default: "circle").}

\item{edge_width_scale}{A scaling factor for edge width.}
}
\value{
A ggplot object
}
\description{
Plot Cell-Cell Communication Network
}
