% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ScaleData}
\alias{ScaleData}
\title{ScaleData}
\usage{
ScaleData(object, features = NULL, assay = "logcounts")
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{features}{selected features to be scaled, default is all features}

\item{assay}{selected assay to be scaled}
}
\value{
an updated SingleCellExperiment object with 'scaled' assay
}
\description{
scale data
}
