% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{sce_merge}
\alias{sce_merge}
\title{sce_merge}
\usage{
sce_merge(sce_list, combineVarParams = list(equiweight = TRUE, ncells = NULL))
}
\arguments{
\item{sce_list}{a named list of SingleCellExperiment objects}

\item{combineVarParams}{parameters that passed to \code{CombineVariableFeatures()}}
}
\value{
A SingleCellExperiment object
}
\description{
merge a list of SingleCellExperiment objects to a single SingleCellExperiment object
}
