% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{subset.SingleCellExperiment}
\alias{subset.SingleCellExperiment}
\title{subset SingleCellExperiment object}
\usage{
\method{subset}{SingleCellExperiment}(x, subset, ...)
}
\arguments{
\item{x}{a SingleCellExperiment object}

\item{subset}{logical expression to subset the object}

\item{...}{additional paramter that will be ignored}
}
\description{
subset SingleCellExperiment object
}
