% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VlnPlot}
\alias{VlnPlot}
\title{VlnPlot}
\usage{
VlnPlot(object, features, ncol = "auto")
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{features}{cell features (e.g., nCounts)}

\item{ncol}{number of columns if multiple features selected}
}
\value{
violin plot
}
\description{
violine plot of selected features
}
