% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{WriteH5AD}
\alias{WriteH5AD}
\title{WriteH5AD}
\usage{
WriteH5AD(object, file, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{file}{Path to the output .h5ad file}

\item{...}{Additional arguments passed to zellkonverter::writeH5AD}
}
\value{
Invisible NULL
}
\description{
Write H5AD file
}
\details{
Wrapper for zellkonverter::writeH5AD to export SingleCellExperiment objects as .h5ad.
}
