% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{BatchRemover}
\alias{BatchRemover}
\title{BatchRemover}
\usage{
BatchRemover(
  sce,
  batch = NULL,
  HVG = NULL,
  nHVG = 5000,
  assay.type = "logcounts",
  PARAM = NULL,
  restrict = NULL,
  correct.all = FALSE,
  name = "batchcorrect"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{batch}{A factor specifying the batch of origin for each cell if only one batch is supplied in \code{...}. This will be ignored if two or more batches are supplied.}

\item{HVG}{user specified high variable genes. If NULL, it will be extracted automatically from the HVG supplied in \code{...}.}

\item{nHVG}{number of HVG. Only works for \code{HVG = NULL} and will be extracted from the info supplied in \code{...} by default.}

\item{assay.type}{specify the assay used in batch correction. If \code{assay.type} is 'counts', it will try to use 'logcounts' or perform log transformation.}

\item{PARAM}{batch correction method, see also \code{batchelor::batchCorrect()}.}

\item{restrict}{A list of length equal to the number of objects in \code{...}. Each entry of the list corresponds to one batch and specifies the cells to use when computing the correction.}

\item{correct.all}{A logical scalar indicating whether to return corrected expression values for all genes, even if \code{subset.row} is set. Used to ensure that the output is of the same dimensionality as the input.}

\item{name}{Integration record id. Defaults to "batchcorrect".}
}
\value{
A SingleCellExperiment object
}
\description{
Batch correction
}
\details{
State-aware batch correction built on \code{batchelor::batchCorrect()}.
}
