% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CellDimPlot}
\alias{CellDimPlot}
\title{CellDimPlot}
\usage{
CellDimPlot(object, group.by = NULL, reduction = NULL, ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{group.by}{cell identity group. If NULL, uses \code{ActiveIdent(object)}.}

\item{reduction}{reduction method. If NULL, uses \code{DefaultReduction(object)}.}

\item{...}{additional parameters passed to \code{ggsc::sc_dim}}
}
\value{
ggplot object
}
\description{
Cell Dimensional Reduction Plot
}
\details{
Dimensional reduction plot that defaults to the active reduction and active identity.
}
