% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DimPlot}
\alias{DimPlot}
\title{DimPlot}
\usage{
DimPlot(object, reduction = NULL, ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{reduction}{reduction method. If NULL, use the active reduction tracked by sclet.}

\item{...}{additional parameters passed to ggsc::sc_dim}
}
\value{
ggplot object
}
\description{
Dimensional reduction plot
}
