% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{FeatureDimPlot}
\alias{FeatureDimPlot}
\title{FeatureDimPlot}
\usage{
FeatureDimPlot(object, features, layer = NULL, reduction = NULL, ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{features}{features to plot}

\item{layer}{layer to extract expression from. If NULL, uses \code{DefaultLayer(object)}.}

\item{reduction}{reduction method. If NULL, uses \code{DefaultReduction(object)}.}

\item{...}{additional parameters passed to \code{ggsc::sc_feature}}
}
\value{
ggplot object
}
\description{
Feature Dimensional Reduction Plot
}
\details{
A wrapper to plot feature expression on dimensional reduction, defaulting to the active reduction and active layer.
}
