% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{FindAllMarkers}
\alias{FindAllMarkers}
\title{Gene expression markers for all identity classes}
\usage{
FindAllMarkers(
  object,
  min.pct = 0.01,
  logfc.threshold = 0.1,
  return.thresh = 0.01,
  only.pos = FALSE,
  base = 2,
  pseudocount.use = 1,
  assay = NULL,
  ...
)
}
\arguments{
\item{object}{A 'SingleCellExperiment' object}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{base}{log base. Default is 2}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 1 by default.}

\item{assay}{assay used for marker testing. If NULL, sclet resolves from
\code{DefaultAssay(object)} and falls back to \code{logcounts} when needed.}

\item{...}{additional parameters}
}
\description{
Find markers for each of the identity classes
}
