% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindClusters}
\alias{FindClusters}
\title{FindClusters}
\usage{
FindClusters(object, resolution = 0.5)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{resolution}{resolution in clustering}
}
\value{
updated SingleCellExperiment object with clustering memberships
}
\description{
identify clusters
}
