% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FindMarkers}
\alias{FindMarkers}
\title{FindMarkers}
\usage{
FindMarkers(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  min.pct = 0.01,
  logfc.threshold = 0.1,
  base = 2,
  pseudocount.use = 1,
  ...
)
}
\arguments{
\item{object}{A 'SingleCellExperiment' object}

\item{ident.1}{Identity class to define markers for}

\item{ident.2}{A second identity class for comparison. If NULL (default) -
use all other cells for comparison.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1}

\item{base}{log base. Default is 2}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 1 by default.}

\item{...}{additional parameters}
}
\description{
Find markers
}
