% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FindVariableFeatures}
\alias{FindVariableFeatures}
\title{FindVariableFeatures}
\usage{
FindVariableFeatures(object, nfeatures = 2000)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{nfeatures}{number of features to be selected as highly variable features}
}
\value{
an updated SingleCellExperiment object with identified highly variable features
}
\description{
identify variable features
}
