% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FindVariableFeatures}
\alias{FindVariableFeatures}
\title{FindVariableFeatures}
\usage{
FindVariableFeatures(object, nfeatures = 2000, method = "scran", ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{nfeatures}{number of features to be selected as highly variable features}

\item{method}{one of 'scran' or 'scrapper'. The legacy value \code{'seurat'}
is accepted for transition but is deprecated and redirected to \code{'scran'}.}

\item{...}{additional parameters for variance modelling. For \code{method = "scran"},
they are passed to \code{scran::modelGeneVar()}. For \code{method = "scrapper"},
function-level arguments such as \code{block}, \code{num.threads}, \code{assay.type},
\code{include.per.block}, \code{more.var.args} and \code{more.choose.args} are supported;
any remaining named arguments are forwarded into \code{more.var.args}.}
}
\value{
an updated SingleCellExperiment object with identified highly variable features
}
\description{
identify variable features
}
