% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sling_plot.R
\name{genecurve_plot}
\alias{genecurve_plot}
\title{Generate gene expression curves across pseudotime}
\usage{
genecurve_plot(
  sce,
  assay_name = "logcounts",
  features,
  lineaegs = NULL,
  pseudotime.data = "slingPseudotime",
  method = c("smooth", "gam"),
  point = FALSE,
  line.size = 0.8,
  alpha = 0.5,
  size = 1,
  ncol = NULL,
  se = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the expression data and pseudotime.}

\item{assay_name}{The assay to use for extracting gene expression data (default: "logcounts").}

\item{features}{A vector of gene names for which to plot expression curves.}

\item{lineaegs}{A vector of cell fates to visualize (optional). If NULL, all available fates will be used.}

\item{pseudotime.data}{The column name in the SingleCellExperiment object containing pseudotime data (default: "slingPseudotime").}

\item{method}{A character vector specifying the method for curve fitting. Options are "smooth" (default) or "gam" (Generalized Additive Model).}

\item{point}{A logical value indicating whether to include points for individual cells (default: FALSE).}

\item{line.size}{Line thickness for the plot (default: 0.8).}

\item{alpha}{Transparency for the points in the plot (default: 0.5).}

\item{size}{Size of the points in the plot (default: 1).}

\item{ncol}{Number of columns to use in facet_wrap (default: NULL).}

\item{se}{A logical value indicating whether to include the standard error for smoothing (default: TRUE).}
}
\value{
A ggplot object displaying gene expression curves across pseudotime.
}
\description{
This function visualizes the expression of specified genes across pseudotime using either
smoothed line plots or GAM (Generalized Additive Models) predictions. It generates a
ggplot showing the expression of selected genes along pseudotime for different cell fates.
}
