% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{GroupHeatmap}
\alias{GroupHeatmap}
\title{GroupHeatmap}
\usage{
GroupHeatmap(object, features, group.by = NULL, layer = NULL, ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{features}{features to plot}

\item{group.by}{cell identity group. If NULL, uses \code{ActiveIdent(object)}.}

\item{layer}{layer to extract expression from. If NULL, uses \code{DefaultLayer(object)}.}

\item{...}{additional parameters passed to underlying plotting function.}
}
\value{
ggplot or ComplexHeatmap object
}
\description{
Group Heatmap Plot
}
\details{
Plot a heatmap of features across cell groups.
If \code{ggsc} handles it, we use it, otherwise fall back to \code{scater}.
}
