% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{has_supercell}
\alias{has_supercell}
\title{has_supercell}
\usage{
has_supercell(object, id = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{id}{optional aggregation record id}
}
\value{
logical
}
\description{
Check whether SuperCell results are available
}
