% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sling_plot.R
\name{lineage_plot}
\alias{lineage_plot}
\alias{plot_lineage}
\title{Visualize Single-Cell Trajectories on the basis of ggsc package}
\usage{
lineage_plot(
  sce,
  dims = c(1, 2),
  group,
  reduction = "UMAP",
  lineages = NULL,
  cells = NULL,
  slot = "data",
  mapping = NULL,
  geom = sc_geom_point,
  linewidth = 0.9,
  ...
)

plot_lineage(
  sce,
  dims = c(1, 2),
  group,
  reduction = "UMAP",
  lineages = NULL,
  cells = NULL,
  slot = "data",
  mapping = NULL,
  geom = sc_geom_point,
  linewidth = 0.9,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the results from Slingshot.}

\item{dims}{A vector specifying the dimensions of the reduction result to plot, default the top 2 dimensions are used.}

\item{group}{The name of the column in colData(sce) containing cluster labels.}

\item{reduction}{The name of the dimensionality reduction to plot. Default is \code{"UMAP"}.}

\item{lineages}{The lineages plotted in this function. Default all lineages are used.}

\item{cells}{A vector of cell names (IDs) to be included in the plot. If \code{NULL}, all cells are plotted.}

\item{slot}{The assay slot to use for dimensionality reduction, default is \code{"data"}. Alternatively, can be set to \code{"logcounts"}, \code{"counts"}, etc.}

\item{mapping}{A mapping for aesthetics such as \code{color}, \code{size}, etc., passed to \code{ggplot2}. By default, \code{NULL} and \code{color} will be mapped automatically.}

\item{geom}{A ggplot2 geometry to use for plotting. Default is \code{sc_geom_point}, which generates scatter points.}

\item{linewidth}{A numeric value indicating the line width for the trajectory paths. Default is 0.9.}

\item{...}{Additional parameters passed to \code{ggplot2} plotting functions or the geometry.}
}
\value{
A ggplot object displaying the UMAP plot with overlaid trajectories.
}
\description{
This function visualizes single-cell trajectories using UMAP (or other specified reduction method) and overlays the trajectory curves.
The cells are colored by their respective lineage trajectories and the lineages are drawn as curves based on Slingshot's inferred trajectories.
}
\examples{
\dontrun{
p <- lineage_plot(sce = pancreas, group = "celltype", reduction = "UMAP")
print(p)
}

}
