% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PercentageFeatureSet}
\alias{PercentageFeatureSet}
\title{PercentageFeatureSet}
\usage{
PercentageFeatureSet(object, pattern = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{pattern}{the pattern to search for}
}
\value{
the percentage of each cell
}
\description{
calculate percentage of genes that matched a pattern
}
