% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PercentageFeatureSet}
\alias{PercentageFeatureSet}
\title{PercentageFeatureSet}
\usage{
PercentageFeatureSet(object, pattern = NULL, feature = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{pattern}{the pattern to search for}

\item{feature}{search pattern from feature of the object. If NULL, use the \code{rownames(object)}, otherwise, use \code{rowData(object)[[feature]]}}
}
\value{
the percentage of each cell
}
\description{
calculate percentage of genes that matched a pattern
}
