% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_heatmap.R
\name{pseudo_heatmap}
\alias{pseudo_heatmap}
\alias{plot_pseudo_heatmap}
\title{Plot gene expression heatmap along pseudotime trajectory}
\usage{
pseudo_heatmap(
  sce,
  features,
  assay_name = "logcounts",
  pseudotime.data = "slingPseudotime",
  lineage = NULL,
  scale = TRUE,
  sort = TRUE,
  cluster_columns = FALSE,
  cluster_rows = FALSE
)

plot_pseudo_heatmap(
  sce,
  features,
  assay_name = "logcounts",
  pseudotime.data = "slingPseudotime",
  lineage = NULL,
  scale = TRUE,
  sort = TRUE,
  cluster_columns = FALSE,
  cluster_rows = FALSE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing gene expression data and pseudotime information.}

\item{features}{A character vector of gene names to be visualized in the heatmap.}

\item{assay_name}{A string specifying the assay to use for gene expression data. Default is \code{"logcounts"}.}

\item{pseudotime.data}{A string specifying the pseudotime data column in \code{colData(sce)}. Default is \code{"slingPseudotime"}.}

\item{lineage}{A string specifying the lineage (cell fate) for the pseudotime trajectory. Default is \code{NULL}, in which case the first column of the pseudotime data is used.}

\item{scale}{A logical indicating whether to scale the gene expression values. Default is \code{TRUE}.}

\item{sort}{A logical indicating whether to sort genes by gradient of expression. Default is \code{TRUE}.}

\item{cluster_columns}{A logical indicating whether to perform hierarchical clustering on cells' pseudotime.  Default is \code{FALSE}.}

\item{cluster_rows}{A logical indicating whether to perform hierarchical clustering on gene expressions.  Default is \code{FALSE}.}
}
\value{
A \code{ComplexHeatmap} object representing the gene expression heatmap.
}
\description{
This function generates a heatmap of gene expression along a given pseudotime trajectory,
with options for scaling, sorting, clustering decision, and lineage selection. It also handles genes with all
\code{NaN} expression values by issuing a warning and removing them from the analysis.
}
