% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sling_plot.R
\name{pseudo_plot}
\alias{pseudo_plot}
\alias{plot_pseudotime}
\title{Visualize pseudotime data on reduction plots}
\usage{
pseudo_plot(
  sce,
  reduction = NULL,
  pseudotime.data = "slingPseudotime",
  lineage = NULL,
  ...
)

plot_pseudotime(
  sce,
  reduction = NULL,
  pseudotime.data = "slingPseudotime",
  lineage = NULL,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object after \code{RunSlingshot()} which contains the pseudotime and reduction data.}

\item{reduction}{A string specifying the reduction method to be used for plotting. Default is \code{NULL}.}

\item{pseudotime.data}{The name of the column in the colData of \code{sce} that contains the pseudotime data. Default is \code{"slingPseudotime"}.}

\item{lineage}{selected lineage to be displayed, all lineages will be used by default.}

\item{...}{Additional parameters to be passed to \code{ggplot2} and other plot settings.}
}
\value{
A \code{ggplot} object representing the pseudotime visualization.
}
\description{
Visualize pseudotime data on reduction plots
}
