% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Read10X}
\alias{Read10X}
\title{Read10X}
\usage{
Read10X(data.dir, ...)
}
\arguments{
\item{data.dir}{directory of the data}

\item{...}{additional parameters passed to 'DropletUtils::read10xCounts()'}
}
\value{
a 'SingleCellExperiment' object
}
\description{
read 10x single cell expression matrix folder
}
