% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Read10X}
\alias{Read10X}
\title{Read10X}
\usage{
Read10X(data.dir, backend = c("memory", "HDF5"), ...)
}
\arguments{
\item{data.dir}{directory of the data}

\item{backend}{backend of the data, default is "memory", support "HDF5"}

\item{...}{additional parameters passed to 'DropletUtils::read10xCounts()'}
}
\value{
a 'SingleCellExperiment' object
}
\description{
read 10x single cell expression matrix folder
}
