% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot_list}
\alias{colData}
\alias{rowData}
\alias{colData<-}
\alias{rowData<-}
\alias{assay}
\alias{plotColData}
\alias{logNormCounts}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aplot}{\code{\link[aplot:plot_list]{plot_list()}}}

  \item{scater}{\code{\link[scater:plotColData]{plotColData()}}}

  \item{scuttle}{\code{\link[scuttle:logNormCounts]{logNormCounts()}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:assay]{assay()}}, \code{\link[SummarizedExperiment:colData]{colData()}}, \code{\link[SummarizedExperiment:colData<-]{colData<-()}}, \code{\link[SummarizedExperiment:rowData]{rowData()}}, \code{\link[SummarizedExperiment:rowData<-]{rowData<-()}}}
}}

