% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{RunCellChat}
\alias{RunCellChat}
\alias{runCellChat}
\title{RunCellChat}
\usage{
RunCellChat(
  sce,
  group = "label",
  assay_name = NULL,
  layer = NULL,
  species = "human",
  db_item = c("Secreted Signaling"),
  type = "triMean",
  trim = 0.1,
  min.cells = 10,
  name = "cellchat",
  return = c("cellchat", "sce", "both")
)

runCellChat(
  sce,
  group = "label",
  assay_name = NULL,
  layer = NULL,
  species = "human",
  db_item = c("Secreted Signaling"),
  type = "triMean",
  trim = 0.1,
  min.cells = 10,
  name = "cellchat",
  return = c("cellchat", "sce", "both")
)
}
\arguments{
\item{sce}{singlecellExperiment object.}

\item{group}{parameter to group data, must be one column in colData(obj).}

\item{assay_name}{the name of assay. If NULL, sclet resolves it from the
selected \code{layer}.}

\item{layer}{layer used for communication analysis. If NULL, use
\code{DefaultLayer(sce)}.}

\item{species}{one of human, mouse}

\item{db_item}{one of  "all", "except Non-protein","Secreted Signaling", "ECM-Receptor", "Cell-Cell Contact" and "Non-protein Signaling"}

\item{type}{one of  "triMean", "truncatedMean", "thresholdedMean", "median"}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before the mean is computed}

\item{min.cells}{min cells to filter}

\item{name}{communication record id. Defaults to \code{"cellchat"}.}

\item{return}{one of \code{"cellchat"}, \code{"sce"} or \code{"both"}}
}
\value{
by default a CellChat object; if \code{return = "sce"}, returns the
updated SingleCellExperiment; if \code{return = "both"}, returns a list with both
}
\description{
wrapper function for running cellchat on default parameter.
}
\details{
https://htmlpreview.github.io/?https://github.com/jinworks/CellChat/blob/master/tutorial/CellChat-vignette.html#c-starting-from-a-singlecellexperiment-object
}
