% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.r
\name{runCellChat}
\alias{runCellChat}
\title{wrapper function for running cellchat on default parameter.}
\usage{
runCellChat(
  sce,
  group = NULL,
  assay_name = "logcounts",
  species = "human",
  db_item = c("Secreted Signaling"),
  type = "triMean",
  trim = 0.1,
  min.cells = 10
)
}
\arguments{
\item{sce}{singlecellExperiment object.}

\item{group}{parameter to group data, must be one column in colData(obj).}

\item{assay_name}{the name of assay}

\item{species}{one of human, mouse}

\item{db_item}{one of  "all", "except Non-protein","Secreted Signaling", "ECM-Receptor", "Cell-Cell Contact" and "Non-protein Signaling"}

\item{type}{one of  "triMean", "truncatedMean", "thresholdedMean", "median"}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before the mean is computed}

\item{min.cells}{min cells to filter}
}
\description{
https://htmlpreview.github.io/?https://github.com/jinworks/CellChat/blob/master/tutorial/CellChat-vignette.html#c-starting-from-a-singlecellexperiment-object
}
