% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.r
\name{RunEnrichment}
\alias{RunEnrichment}
\title{RunEnrichment}
\usage{
RunEnrichment(
  markers,
  db = c("GO", "KEGG"),
  orgDb = "org.Hs.eg.db",
  keyType = "SYMBOL",
  ont = "BP",
  ...
)
}
\arguments{
\item{markers}{Data frame returned by FindAllMarkers or FindMarkers}

\item{db}{Database to use ("GO" or "KEGG")}

\item{orgDb}{OrgDb object (e.g., org.Hs.eg.db)}

\item{keyType}{Key type of gene IDs (e.g., "SYMBOL", "ENTREZID")}

\item{ont}{Ontology for GO ("BP", "MF", "CC", "ALL")}

\item{...}{Additional arguments passed to enrichGO or enrichKEGG}
}
\value{
A clusterProfiler result object (compareClusterResult)
}
\description{
Run Functional Enrichment Analysis
}
\details{
Wrapper for clusterProfiler enrichment analysis on marker genes.
}
