% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{RunPCA}
\alias{RunPCA}
\title{RunPCA}
\usage{
RunPCA(
  object,
  subset_row = NULL,
  exprs_values = NULL,
  layer = NULL,
  ncomponents = 50,
  ...
)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{subset_row}{subset of rows used for PCA}

\item{exprs_values}{assay used for PCA. If NULL, sclet resolves it from the
selected \code{layer}.}

\item{layer}{layer used for PCA. If NULL, use \code{DefaultLayer(object)}.}

\item{ncomponents}{number of components}

\item{...}{additional parameters passed to 'scater::runPCA'}
}
\value{
an updated SingleCellExperiment object with PCA dimension reduction
}
\description{
Run PCA
}
