% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  object,
  subset_row = NULL,
  exprs_values = "logcounts",
  ncomponents = 50,
  ...
)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{subset_row}{subset of rows used for PCA}

\item{exprs_values}{assay used for PCA}

\item{ncomponents}{number of components}

\item{...}{additional parameters passed to 'scater::runPCA'}
}
\value{
an updated SingleCellExperiment object with PCA dimension reduction
}
\description{
run pca
}
