% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC.R
\name{runSCENIC}
\alias{runSCENIC}
\title{wrapper function for running SCENIC on default parameter.}
\usage{
runSCENIC(
  sce,
  species,
  nCores,
  dbDir,
  assay_name = "logcounts",
  minCountsPerGene = 3,
  minSamples = 0.01
)
}
\arguments{
\item{sce}{A singlecellExperiment object.}

\item{species}{The species of the data, one of human, mouse or drosophila_melanogaster.}

\item{nCores}{The number of cores to use for parallel processing.}

\item{dbDir}{The directory path to the cisTarget databases files.}

\item{assay_name}{the name of assay.}

\item{minCountsPerGene}{The minimum number of counts per gene required for inclusion in the analysis.}

\item{minSamples}{The minimum proportion of samples in which a gene must be detected to be included in the analysis.}
}
\value{
A matrix of regulon activity scores for each cell.
}
\description{
https://htmlpreview.github.io/?https://github.com/aertslab/SCENIC/blob/master/inst/doc/SCENIC_Running.html
}
