% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC.R
\name{runSCENIC}
\alias{runSCENIC}
\title{wrapper function for running SCENIC on default parameter.}
\usage{
runSCENIC(sce, species, nCores, dbDir, assay_name = "counts")
}
\arguments{
\item{sce}{A singlecellExperiment object.}

\item{species}{The species of the data, one of human, mouse or drosophila_melanogaster.}

\item{nCores}{The number of cores to use for parallel processing.}

\item{dbDir}{The directory path to the cisTarget databases files.}

\item{assay_name}{the name of assay.}
}
\value{
The class contains the options/settings for a run of SCENIC.
}
\description{
https://htmlpreview.github.io/?https://github.com/aertslab/SCENIC/blob/master/inst/doc/SCENIC_Running.html
}
