% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{RunSingleR}
\alias{RunSingleR}
\title{RunSingleR}
\usage{
RunSingleR(object, ref = NULL, labels = NULL, assay.type = "logcounts", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{ref}{Reference dataset (e.g. from celldex). If NULL, will try to use HumanPrimaryCellAtlasData.}

\item{labels}{Labels in the reference dataset. If NULL, tries to use 'label.main'.}

\item{assay.type}{Assay to use for annotation. Default is "logcounts".}

\item{...}{Additional arguments passed to SingleR::SingleR}
}
\value{
A SingleCellExperiment object with annotation added to colData
}
\description{
Run SingleR for cell type annotation
}
\details{
Wrapper for SingleR to perform automatic cell type annotation.
}
