% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCell.R
\name{RunSuperCell}
\alias{RunSuperCell}
\title{RunSuperCell}
\usage{
RunSuperCell(
  object,
  assay = "logcounts",
  layer = NULL,
  nHVG = 2000,
  hvg_method = "scran",
  cellname = "Barcode",
  gamma = 20,
  k.knn = 5,
  name = "supercell"
)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{assay}{selected assay used to build KNN graph. If \code{layer} is provided,
this is treated as a legacy argument alias.}

\item{layer}{layer used to build the metacell representation. If \code{NULL},
sclet resolves it from \code{DefaultLayer(object)}.}

\item{nHVG}{number of HVGs to use}

\item{hvg_method}{one of \code{'scran'}, \code{'scrapper'}, or legacy \code{'seurat'};
defaults to the Bioconductor-native \code{'scran'} workflow. See also
\code{FindVariableFeatures()}.}

\item{cellname}{IDs to label the cells, use 'Barcode' by default}

\item{gamma}{graining level, that is number_of_cells / number_of_metacells}

\item{k.knn}{number of nearest neighbors to build KNN graph}

\item{name}{aggregation record id. Defaults to \code{"supercell"}.}
}
\value{
SingleCellExperiment object
}
\description{
Run SuperCell aggregation
}
