% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{RunUMAP}
\alias{RunUMAP}
\title{RunUMAP}
\usage{
RunUMAP(object, dims)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{dims}{dimensions used in UMAP}
}
\value{
updated SingleCellExperiment object with UMAP dimension reduction
}
\description{
run umap
}
