% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cci.R
\name{sc_cci_bubble}
\alias{sc_cci_bubble}
\title{sc_cci_bubble}
\usage{
sc_cci_bubble(
  object,
  sources = NULL,
  targets = NULL,
  id = NULL,
  pval_thresh = 0.05
)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{sources}{A character vector of source groups to include. If \code{NULL}, all are included.}

\item{targets}{A character vector of target groups to include. If \code{NULL}, all are included.}

\item{id}{The communication record id to use.}

\item{pval_thresh}{The p-value threshold for filtering interactions.}
}
\value{
A ggplot object
}
\description{
sc_cci_bubble
}
