% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclet-package.R
\docType{package}
\name{sclet-package}
\alias{sclet}
\alias{sclet-package}
\alias{package-sclet}
\title{sclet: A Lightweight Toolkit for Single-Cell Data Analysis}
\description{
Help students migrate from Seurat to SingleCellExperiment.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Kaijia Huang \email{2947569391@qq.com} [contributor]
  \item Nan He \email{hinna@i.smu.edu.cn} [contributor]
}

}
\keyword{internal}
