% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclet-package.R
\docType{package}
\name{sclet-package}
\alias{sclet}
\alias{sclet-package}
\alias{package-sclet}
\title{sclet: Lightweight Single-Cell Analysis Toolkit for SingleCellExperiment}
\description{
Provides a lightweight set of Seurat-like helpers for single-cell analysis based on the Bioconductor SingleCellExperiment class. Includes preprocessing, dimensionality reduction, clustering, visualization, batch correction, and pseudobulk differential expression, plus interoperability with AnnData. Optional wrappers integrate popular downstream tools for trajectory inference, enrichment analysis, cell type annotation, cell-cell communication, and differential abundance.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Kaijia Huang \email{2947569391@qq.com} [contributor]
  \item Nan He \email{hinna@i.smu.edu.cn} [contributor]
  \item Ming Li \email{limiang929@gmail.com} [contributor]
  \item Junrui Li \email{1523467190@qq.com} [contributor]
}

}
\keyword{internal}
