% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{VariableFeaturePlot}
\alias{VariableFeaturePlot}
\title{VariableFeaturePlot}
\usage{
VariableFeaturePlot(object, label = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{label}{display selected features}
}
\value{
scatter plot
}
\description{
variable feature plot
}
