% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VariableFeaturePlot}
\alias{VariableFeaturePlot}
\title{VariableFeaturePlot}
\usage{
VariableFeaturePlot(object, label = NULL, method = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{label}{display selected features}

\item{method}{one of 'seurat' or 'scran'. If NULL, use stored method or infer from rowData.}
}
\value{
scatter plot
}
\description{
variable feature plot
}
