% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{VariableFeatures}
\alias{VariableFeatures}
\title{VariableFeatures}
\usage{
VariableFeatures(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
highly variable features
}
\description{
get variable features
}
