% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{VariableFeatures}
\alias{VariableFeatures}
\title{VariableFeatures}
\usage{
VariableFeatures(object, method = "scran", ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{method}{one of 'seurat', 'scran' or 'scrapper'}

\item{...}{additional parameters for ranking the \code{method = "scran"} HVG table}
}
\value{
highly variable features
}
\description{
get variable features
}
