% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{VariableFeatures}
\alias{VariableFeatures}
\title{VariableFeatures}
\usage{
VariableFeatures(object, method = "scran", ...)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{method}{one of 'seurat' or 'scran'}

\item{...}{additional parameters for 'method = "scran"', see also \code{scran::getTopHVGs()}}
}
\value{
highly variable features
}
\description{
get variable features
}
