% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{VelocityPlot}
\alias{VelocityPlot}
\title{Plot RNA Velocity}
\usage{
VelocityPlot(sce, reduction = "UMAP", group.by = NULL, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object with velocity calculated.}

\item{reduction}{String specifying the reduction to plot on. Defaults to "UMAP".}

\item{group.by}{String specifying the colData column for cell grouping. If NULL,
uses \code{ActiveIdent(sce)}, including the special \code{"colLabels"} identity source.}

\item{...}{Additional arguments passed to plotting function.}
}
\value{
A ggplot object.
}
\description{
Plots RNA velocity stream or arrows on a reduced dimension plot.
}
